SO... MANY... CHANGES...

+++ src/com/squidkingdom/Gavel/FinalCompare.java
@@ -10,8 +10,8 @@ class FinalsCompare implements Comparator<Team>{
             if (m1.getSpeaks() < m2.getSpeaks()) return -1;
             if (m1.getSpeaks() > m2.getSpeaks()) return 1;
             else {
-                float opp1 = m1.opp[0].totalSpeaks + m1.opp[1].totalSpeaks + m1.opp[2].totalSpeaks + m1.opp[3].totalSpeaks + m1.opp[4].totalSpeaks;
-                float opp2 = m2.opp[0].totalSpeaks + m2.opp[1].totalSpeaks + m2.opp[2].totalSpeaks + m2.opp[3].totalSpeaks + m2.opp[4].totalSpeaks;
+                double opp1 = m1.opp[0].totalSpeaks + m1.opp[1].totalSpeaks + m1.opp[2].totalSpeaks + m1.opp[3].totalSpeaks + m1.opp[4].totalSpeaks;
+                double opp2 = m2.opp[0].totalSpeaks + m2.opp[1].totalSpeaks + m2.opp[2].totalSpeaks + m2.opp[3].totalSpeaks + m2.opp[4].totalSpeaks;
                 if (opp1 < opp2) return -1;
                 if (opp1 > opp2) return 1;
                 else return 0;
